+++ src/nns/leaf/options/list.rs
@@ -4,9 +4,10 @@ use crate::{
     nns::{
         NnsCommandError,
         leaf::{
-            commands::{VERBOSE_ARG, list_command, list_usage, parse_leaf_matches},
+            commands::{VERBOSE_ARG, list_command, list_usage},
             model::NnsLeafCommandSpec,
         },
+        parse_nns_matches,
     },
 };
 use std::ffi::OsString;
@@ -28,10 +29,9 @@ impl NnsLeafListOptions {
     where
         I: IntoIterator<Item = OsString>,
     {
-        let matches =
-            parse_leaf_matches(list_command(spec, default_source_endpoint), args, || {
-                list_usage(spec, default_source_endpoint)
-            })?;
+        let matches = parse_nns_matches(list_command(spec, default_source_endpoint), args, || {
+            list_usage(spec, default_source_endpoint)
+        })?;
         let common = NnsCommonOptions::from_matches(&matches);
         Ok(Self {
             network: common.network,