+++ src/sns/commands/options/mod.rs
@@ -1,3 +1,4 @@
+mod common;
 mod list;
 mod lookup;
 mod neurons;